+++ NEWS.md
@@ -112,3 +112,4 @@
 
 # Racmacs 1.1.26
 * Add option to select the algorithm MASS::kde2d or ks::kde for calculating 2D density blobs.
+* Add functions to help plot and inspect bootstrap blobs, `agBootstrapBlob()`, `srBootstrapBlob()`, `agBootstrapCoords()`, `srBootstrapCoords()`, `blob()`